+++ runner/.gitignore
@@ -1,3 +1,3 @@
-subproc
-runner
+sonar-mock
+sonar-daemon-runner
 *.o
