+++ runner/sonar-daemon-runner.c
